# docs/swagger.yaml
basePath: /api
definitions:
  courseerror.CourseError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  entity.AdminCredentials:
    properties:
      code:
        type: string
      login:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  entity.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:70
info:
  contact: {}
  title: Приложение course
  version: "1.0"
paths:
  /v1/admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: логин, пароль, код подтверждения
        in: body
        name: adminData
        required: true
        schema:
          $ref: '#/definitions/entity.AdminCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или декодирование сообщения
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Неправильный логин, пароль или код
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Администратор не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Залогиниться администратору
      tags:
      - Методы для администрирования
...
  /v1/admin/management/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: логин, пароль
        in: body
        name: adminData
        required: true
        schema:
          $ref: '#/definitions/entity.AdminCredentials'
      responses:
        "200":
          description: image/png
          schema:
            type: string
        "400":
          description: Провалена валидация или декодирование сообщения
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Нет прав
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "409":
          description: Невозможно создать админа
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Создать профиль нового администратора
      tags:
      - Методы для администрирования
...
  /v1/admin/management/resetKey:
    patch:
      parameters:
      - description: логин
        in: query
        name: login
        required: true
        type: string
      responses:
        "200":
          description: image/png
          schema:
            type: string
        "400":
          description: Провалена валидация или не получилось декодировать сообщение
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Администратор не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить ключ администратора
      tags:
      - Методы для администрирования
  /v1/admin/management/resetPassword:
    patch:
      consumes:
      - application/json
      parameters:
      - description: логин, пароль
        in: body
        name: adminData
        required: true
        schema:
          $ref: '#/definitions/entity.AdminCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или декодирование сообщения
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Неправильный логин, пароль или код
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Администратор не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Изменить пароль администратора
      tags:
      - Методы для администрирования
...
  /v1/admin/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: логин, пароль, код подтверждения
        in: body
        name: adminData
        required: true
        schema:
          $ref: '#/definitions/entity.AdminCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или декодирование сообщения
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Неверный код или пара логин-пароль
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Код не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Верифицировать аутентификатор
      tags:
      - Методы для администрирования
...
securityDefinitions:
  AdminApiKeyAuth:
    in: cookie
    name: admin_auth
    type: apiKey
  ApiKeyAuth:
    in: cookie
    name: auth
    type: apiKey
swagger: "2.0"